add waivers to event routes
