+++ src/components/events/createEvent/EventsAdditionalInfo.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { useFormContext, Controller } from 'react-hook-form';
 import { Form, Input, Upload, Button } from 'antd';
 import { RightOutlined } from '@ant-design/icons';
+import PropTypes from 'prop-types';
 
-const EventsAdditionalInfo = () => {
+const EventsAdditionalInfo = ({ setFilesList }) => {
   const { control } = useFormContext();
 
   return (
@@ -23,29 +24,31 @@ const EventsAdditionalInfo = () => {
           </Form.Item>
         )}
       />
-      <Controller
-        control={control}
-        name="fileAttachments"
-        render={({ field: { onChange, ref } }) => (
-          <Form.Item label="Upload Forms">
-            <Upload multiple ref={ref} onChange={onChange}>
-              <Button
-                icon={<RightOutlined />}
-                style={{
-                  background: 'rgba(108, 194, 74, 0.25)',
-                  color: 'rgba(0, 0, 0, 0.85)',
-                  border: 'rgba(17, 87, 64, 0.25)',
-                }}
-              >
-                {' '}
-                Click to Upload
-              </Button>
-            </Upload>
-          </Form.Item>
-        )}
-      />
+      <Form.Item label="Upload Forms">
+        <Upload
+          multiple
+          onChange={e => setFilesList(e.fileList)}
+          customRequest={e => e.onSuccess('Ok')}
+        >
+          <Button
+            icon={<RightOutlined />}
+            style={{
+              background: 'rgba(108, 194, 74, 0.25)',
+              color: 'rgba(0, 0, 0, 0.85)',
+              border: 'rgba(17, 87, 64, 0.25)',
+            }}
+          >
+            {' '}
+            Click to Upload
+          </Button>
+        </Upload>
+      </Form.Item>
     </div>
   );
+};
+
+EventsAdditionalInfo.propTypes = {
+  setFilesList: PropTypes.func.isRequired,
 };
 
 export default EventsAdditionalInfo;
